Fix: Import and type errors
