+++ frontend/src/hooks/UseVideos.ts
@@ -24,8 +24,6 @@ export const useVideos = (folderPath: string) => {
   useEffect(() => {
     const fetchVideos = async () => {
       try {
-        console.log("Fetching videos from folder:", folderPath);
-
         // Fetch video paths using invoke
         const response: ResponseData = await invoke(
           "get_all_videos_with_cache",
@@ -43,7 +41,6 @@ export const useVideos = (folderPath: string) => {
 
         const videoUrls: VideoData[] = [];
 
-        // Iterate through each year in the response
         for (const year in response) {
           if (
             !response.hasOwnProperty(year) ||
@@ -107,89 +104,3 @@ export const useVideos = (folderPath: string) => {
   return { videos, loading };
 };
 
-interface AddMultipleImagesResult {
-  data: any | null;
-  error: string | null;
-  isLoading: boolean;
-}
-
-export function useAddMultipleImages() {
-  const [result, setResult] = useState<AddMultipleImagesResult>({
-    data: null,
-    error: null,
-    isLoading: false,
-  });
-
-  const addMultipleImages = async (paths: string[]): Promise<void> => {
-    setResult({ data: null, error: null, isLoading: true });
-
-    try {
-      const response = await fetch(
-        "http://localhost:8000/images/multiple-images",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ paths }),
-        }
-      );
-      console.log(response);
-      if (!response.ok) {
-        throw new Error("Failed to add multiple images");
-      }
-
-      const data = await response.json();
-      setResult({ data, error: null, isLoading: false });
-    } catch (error) {
-      setResult({
-        data: null,
-        error: (error as Error).message,
-        isLoading: false,
-      });
-    }
-  };
-
-  return { addMultipleImages, ...result };
-}
-
-interface Image {
-  id: string;
-  path: string;
-  // Add other properties as needed
-}
-
-export function useFetchAllImages() {
-  const [images, setImages] = useState<Image[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const fetchImages = async () => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const response = await fetch("http://127.0.0.1:8000/images/all-images", {
-        headers: {
-          accept: "application/json",
-        },
-      });
-
-      if (!response.ok) {
-        throw new Error("Failed to fetch images");
-      }
-      const data = await response.json();
-      console.log(data);
-      setImages(data);
-    } catch (err) {
-      setError((err as Error).message);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchImages();
-  }, []);
-
-  return { images, isLoading, error, refetch: fetchImages };
-}
